+++ lua/terrortown/lang/pl.lua
@@ -130,8 +130,6 @@ L.xfer_received = "{player} dał ci {num} kredyt(ów)."
 
 -- Radio tab in equipment menu
 L.radio_name = "Radio"
-L.radio_help = "Klkinij przycisk, by twoje radio zaczeło grać."
-L.radio_notplaced = "Musisz postawić radio, by móc puścić jakieś dźwięki."
 
 -- Radio soundboard buttons
 L.radio_button_scream = "Krzyk"
@@ -2190,3 +2188,23 @@ L.decoy_help_primary = "Rozstaw Wabik"
 
 -- 2024-02-14
 --L.throw_no_room = "You have no space here to throw this device"
+
+-- 2024-03-04
+--L.use_entity = "Press [{usekey}] to use"
+
+-- 2024-03-06
+--L.submenu_gameplay_sounds_title = "Client-Sounds"
+
+--L.header_sounds_settings = "UI Sound Settings"
+
+--L.help_enable_sound_interact = "Interaction sounds are those sounds that are played when opening an UI. Such a sound is played for example when interacting with the radio marker."
+--L.help_enable_sound_buttons = "Button sounds are clicky sounds that are played when clicking a button."
+--L.help_enable_sound_message = "Message or notification sounds are played for chat mesages and notifications. They can be quite obnoxious."
+
+--L.label_enable_sound_interact = "Enable interaction sounds"
+--L.label_enable_sound_buttons = "Enable button sounds"
+--L.label_enable_sound_message = "Enable message sounds"
+
+--L.label_level_sound_interact = "Interaction sound level multiplier"
+--L.label_level_sound_buttons = "Button sound level multiplier"
+--L.label_level_sound_message = "Message sound level multiplier"