+++ CHANGELOG.md
@@ -9,10 +9,25 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Added hook ENTITY:ClientUse(), which is triggered clientside if an entity is used
   - Return true to prevent also using this on the server for clientside only usecases
 - Added upstream content files to base TTT2
+- Added `plymeta:IsFullySignedOn()` to allow excluding players that have not gotten control yet (by @EntranceJew)
+- Added hook ENTITY:RemoteUse(ply), which is shared
+  - Return true if only clientside should be used
+- Added RemoteUse to radio, you can now directly access it via use button on marker focus
+- Added sounds to multiple UI interactions (can be disabled in settings: Gameplay > Client-Sounds)
+- Added a globally audible sound when searching a body
+- Added the option to add a subtitle to a marker vision element
 
 ### Changed
 
+- TargetID is now hidden when a marker vision element is focused
+
 ### Fixed
+
+- Fixed the AFK timer accumulating while player not fully joined (by @EntranceJew)
+- 
+### Removed
+
+- Removed radio tab in shop UI
 
 ## [v0.13.1b](https://github.com/TTT-2/TTT2/tree/v0.13.1b) (2024-02-27)